Completed PATCH + DELETE methods

+++ backend/app.py
@@ -39,22 +39,22 @@ def home():
 @app.route("/signup", methods=['POST'])
 def user_signup() -> Tuple[Response, int]:
     data = request.get_json()
-    if not data['name']:
+    if not data.get('name'):
         return (
             jsonify({'error': 'Name is required'}),
             400
         )
-    if not data['email']:
+    if not data.get('email'):
         return (
             jsonify({'error': 'Email is required'}),
             400
         )
-    if not data['password']:
+    if not data.get('password'):
         return (
             jsonify({'error': 'Password is required'}),
             400
         )
-    if not data['confirm_password']:
+    if not data.get('confirm_password'):
         return (
             jsonify({'error': 'Password confirmation is required'}),
             400
@@ -139,11 +139,6 @@ def login() -> Tuple[Response, int]:
             400
         ) 
 
-
-@app.route("/health")
-def health():
-    return {"status": "healthy"}
-
 @app.route("/user/dashboard", methods=['GET'])
 @Authenticator.authenticate_session
 def load_user_dashboard() -> Tuple[Response, int]:
@@ -182,8 +177,6 @@ def load_user_dashboard() -> Tuple[Response, int]:
             project_dict['features'] = feature_list
             dashboard_data.append(project_dict)
 
-                
-
         return (    
             jsonify({"projects": dashboard_data}),
             200
@@ -197,11 +190,11 @@ def load_user_dashboard() -> Tuple[Response, int]:
 
 @app.route("/projects", methods=['POST'])
 @Authenticator.authenticate_session
-def create_new_project() -> Tuple[Response, int]:
+def add_project() -> Tuple[Response, int]:
     data = request.get_json()
-
+    user_id= session['user_id']
     try:
-        user_id = session['user_id']
+        
         project = Project(
             name=data['name'],
             description=data.get('description'),
@@ -211,7 +204,7 @@ def create_new_project() -> Tuple[Response, int]:
         g.db.add(project)
         g.db.flush()
         return (
-            jsonify({'id': project.id, 'parent_userid': user_id, 'name': project.name, 'description': project.description}), 
+            jsonify({'project': {'id': project.id, 'parent_userid': user_id, 'name': project.name, 'description': project.description}}), 
             201
             )
     except Exception as e:
@@ -220,69 +213,72 @@ def create_new_project() -> Tuple[Response, int]:
             400
         )
 
-@app.route("/projects/<int:project_id>", methods=['DELETE'])
+@app.route("/projects/<int:project_id>", methods=['PATCH','DELETE'])
 @Authenticator.authenticate_session
-def delete_project(project_id) -> Tuple[Response, int]:
+def handle_project(project_id) -> Tuple[Response, int]:
+    user_id=session['user_id']
     try:
-        user_id=session['user_id']
         project = g.db.query(Project).filter(Project.id == project_id, Project.parent_userid == user_id).first()
         if not project:
             return (
                 jsonify({"error": "project_id not found"}),
                 404
             )
-        project_name = project.name
-        g.db.delete(project)
-        g.db.flush()
-        return (
-            jsonify({'message': f"{project_name} deleted"}),
-            200
-        )
+        if request.method == 'DELETE':
+
+            project_name = project.name
+            g.db.delete(project)
+            g.db.flush()
+            return (
+                jsonify({'msg': f"{project_name} deleted"}),
+                200
+            )
+        
+        elif request.method == 'PATCH':
+            data = request.get_json()
+            project.name = data.get('name') if data.get('name') is not None else project.name
+            project.description = data.get('description') if data.get('description') is not None else project.description
+            g.db.flush()
+            return (
+                jsonify({"project": {"id": project.id, "name": project.name, "description": project.description}}),
+                200
+            )
+        else:
+            return (
+                jsonify({"error": "invalid http method"}),
+                400
+            )
+        
     except Exception as e:
         return (
             jsonify({"error": str(e)}),
             400
         )
 
-@app.route("/projects/<int:project_id>/features", methods=['GET'])
-@Authenticator.authenticate_session
-def get_project_features(project_id) -> Tuple[Response, int]:
-    try:
-        user_id = session['user_id']
-        project = g.db.query(Project).filter(Project.id == project_id).filter(Project.parent_userid == user_id).first()
-        feature_list = [{'id':f.id, 'name':f.name, 'description':f.description} for f in project.feature_list]
-        if not project:
-            return (
-                jsonify({"error": "project_id not found"}),
-                404
-            )
-        return (
-            jsonify({'project_id': project.id, 'project_name': project.name, 'feature_list': feature_list}),
-            200
-        )
-    except Exception as e:
-        return (
-            jsonify({'error': str(e)}),
-            400
-        )
 
 @app.route("/projects/<int:project_id>/features", methods=['POST'])
 @Authenticator.authenticate_session
-def add_feature_to_project(project_id) -> Tuple[Response, int]:
+def add_feature(project_id) -> Tuple[Response, int]:
     data = request.get_json()
+    user_id = session['user_id']
+    project = g.db.query(Project).filter(Project.parent_userid ==  user_id, Project.id == project_id).first()
 
+    if not project:
+        return (
+            jsonify({"error": "project not found"}),
+            404
+        )
+    
     try:
-
         feature = Feature(
             name=data['name'],
             project_id=project_id,
             description=data.get('description'),
-            priority=data.get('priority'),
         )
         g.db.add(feature)
         g.db.flush()
         return (
-            jsonify({'msg': f'Feature {feature.name} added to project {feature.parent_project.name}'}),
+            jsonify({'feature': {'id':feature.id, 'project_id': feature.project_id, 'name': feature.name, 'description': feature.description}}),
             200
         )
     except Exception as e:
@@ -291,5 +287,196 @@ def add_feature_to_project(project_id) -> Tuple[Response, int]:
             400
         )
 
+@app.route("/features/<int:feature_id>", methods=['DELETE', 'PATCH'])
+@Authenticator.authenticate_session
+def handle_feature(feature_id) -> Tuple[Response, int]:
+    user_id = session['user_id']
+    try:
+        feature = g.db.query(Feature).join(Project).filter(Feature.id == feature_id, Project.parent_userid == user_id).first()
+        
+        if not feature:
+            return (
+                jsonify({'error': 'feature not found'}),
+                404
+            )
+        
+        if request.method == 'DELETE':
+            g.db.delete(feature)
+            g.db.flush()
+
+            return (
+                jsonify({'message': 'feature deleted'}),
+                200
+            )
+        
+        elif request.method == 'PATCH':
+            data = request.get_json()
+            feature.name = data['name'] if data.get('name') is not None else feature.name
+            feature.description = data['description'] if data.get('description') is not None else feature.description
+            g.db.flush()
+            return (
+                jsonify({'feature': {'id':feature.id, 'project_id': feature.project_id, 'name': feature.name, 'description': feature.description}}),
+                200
+            )
+        else:
+            return (
+                jsonify({"error": "invalid http method"}),
+                400
+            )
+    except Exception as e:
+        return (
+            jsonify({'error': str(e)}),
+            400
+        )
+
+
+
+@app.route("/features/<int:feature_id>/tasks", methods=['POST'])
+@Authenticator.authenticate_session
+def add_task(feature_id) -> Tuple[Response, int]:
+    data = request.get_json()
+    user_id = session['user_id']
+    feature = g.db.query(Feature).join(Project).filter(Feature.id == feature_id, Project.parent_userid == user_id).first()
+
+    if not feature:
+        return (
+            jsonify({"error": "feature not found"}),
+            404
+        )
+
+    try:
+        task = Task(
+            name=data['name'],
+            feature_id=feature_id,
+            description=data.get('description'),
+            points=data.get('points')
+        )
+
+        g.db.add(task)
+        g.db.flush()
+        return (
+            jsonify({'task': {'id':task.id, 'feature_id': task.feature_id, 'name': task.name, 'description': task.description}}),
+            201
+        )
+    except Exception as e:
+        return (
+            jsonify({'err': str(e)}),
+            400
+        )
+
+@app.route("/tasks/<int:task_id>", methods=['DELETE', 'PATCH'])
+@Authenticator.authenticate_session
+def handle_task(task_id) -> Tuple[Response, int]:
+    user_id = session['user_id']
+    try:
+        task = g.db.query(Task).join(Feature).join(Project).filter(Task.id == task_id, Project.parent_userid == user_id).first()
+
+        if not task:
+            return (
+                jsonify({"error": "task not found"}),
+                404
+            )
+        
+        if request.method == 'DELETE':
+            g.db.delete(task)
+            g.db.flush()
+            return (
+                jsonify({"message": "task deleted"}),
+                200
+            )
+        elif request.method == 'PATCH':
+            data = request.get_json()
+            task.name = data['name'] if data.get('name') is not None else task.name
+            task.description = data['description'] if data.get('description') is not None else task.description
+            task.points = data['points'] if data.get('points') is not None else task.points
+            task.completed = data['completed'] if data.get('completed') is not None else task.completed
+
+            g.db.flush()
+
+            return (
+                jsonify({'task': {'id':task.id, 'feature_id': task.feature_id, 'name': task.name, 'description': task.description}}),
+                200
+            )
+        else:
+            return (
+                jsonify({"error": "invalid http method"}),
+                400
+            )
+    except Exception as e:
+        return (
+            jsonify({'err': str(e)}),
+            400
+        )
+@app.route("/tasks/<int:task_id>/notes", methods=['POST'])
+@Authenticator.authenticate_session
+def add_note(task_id) -> Tuple[Response, int]:
+    data = request.get_json()
+    user_id = session['user_id']
+    task = g.db.query(Task).join(Feature).join(Project).filter(Task.id == task_id, Project.parent_userid == user_id).first()
+    
+    if not task:
+        return (
+            jsonify({'error': 'task not found'}),
+            404
+        )
+    
+    try:
+        note = Note(
+            task_id=task_id,
+            content=data["content"]
+        )
+
+        g.db.add(note)
+        g.db.flush()
+        return (
+            jsonify({"note": {"id": note.id, "task_id": task_id, "content": note.content}}),
+            201
+        )
+    
+    except Exception as e:
+        return (
+            jsonify({"error": str(e)}),
+            400
+        )
+    
+@app.route("/notes/<int:note_id>", methods=['PATCH', 'DELETE'])
+@Authenticator.authenticate_session
+def handle_note(note_id) -> Tuple[Response, int]:
+    user_id = session['user_id']
+    try:
+        note = g.db.query(Note).join(Task).join(Feature).join(Project).filter(Note.id == note_id, Project.parent_userid == user_id).first()
+
+        if not note:
+            return (
+                jsonify({"error": "note not found"}),
+                404
+            )
+        
+        if request.method == 'DELETE':
+            g.db.delete(note)
+            g.db.flush()
+            return (
+                jsonify({"message": "note deleted"}),
+                200
+            )
+        elif request.method == 'PATCH':
+            data = request.get_json()
+            note.content = data['content'] if data.get('content') is not None else note.content
+            g.db.flush()
+            return (
+                jsonify({"note": {"id": note.id, "task_id": note.task_id, "content": note.content}}),
+                200
+            )
+        else:
+            return (
+                jsonify({"error": "invalid http method"}),
+                400
+            )
+    except Exception as e:
+        return (
+            jsonify({'err': str(e)}),
+            400
+        )
+    
 if __name__ == "__main__":
     app.run(debug=True, port=5001)
